reset-to-issue1
changing the code to reflect issue 1
